Add files via upload
Import layoffs.csv into MySQL, open the attached file, and practice with it in a new file. Reading won't help. nearly 200 lines of advanced techniques are present to clean any sort of data. After learning from this you will be experienced enough to solve intermediate to hard level questions on LeetCode and in real life as well.